Renders names of instructors
