+++ package.json
@@ -7,7 +7,9 @@
     "node": ">=6.13.0"
   },
   "scripts": {
-    "seed": "node database/index.js"
+    "seed": "node database/seed.js",
+    "react-dev": "webpack -d --watch",
+    "server-dev": "nodemon server/index.js"
   },
   "dependencies": {
     "bluebird": "^3.5.2",
@@ -15,17 +17,18 @@
     "faker": "^4.1.0",
     "mysql": "^2.16.0",
     "react": "^16.5.2",
-    "react-dom": "^16.5.2"
+    "react-dom": "^16.5.2",
+    "sequelize": "^4.41.0"
   },
   "devDependencies": {
-    "eslint": "^5.7.0",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.14.0",
     "@babel/core": "^7.1.2",
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.4",
-    "webpack": "^4.21.0",
+    "eslint": "^5.7.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "webpack": "^4.23.1",
     "webpack-cli": "^3.1.2"
   },
   "version": "1.0.0",
