+++ webpack.config.js
@@ -1,0 +1,21 @@
+module.exports = {
+  entry: __dirname + '/client/src/index.jsx',
+  module: {
+    rules: [
+      { 
+        test: [/\.jsx$/],
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-react', '@babel/preset-env']
+          }
+        }
+      }
+    ]
+  },
+   output: {
+    filename: 'bundle.js',
+    path: __dirname + '/client/dist'
+  }
+};
